feature(frontend): Adding Analyzer empty state
